controllers/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\github\MemeMarketCap\controllers\cli\cli.py: 1
# C:\github\MemeMarketCap\controllers\cli\commands\cmd_generate.py: 1
# C:\github\MemeMarketCap\controllers\cli\commands\cmd_imgflip.py: 1
# C:\github\MemeMarketCap\controllers\cli\commands\cmd_reddit.py: 1
# C:\github\MemeMarketCap\controllers\cli\commands\cmd_tasks.py: 1
Click == 7.0

# C:\github\MemeMarketCap\controllers\controller\generators\classes\settings.py: 5
# C:\github\MemeMarketCap\controllers\controller\generators\classes\simulator.py: 10
Faker == 4.0.2

# C:\github\MemeMarketCap\controllers\controller\api\database.py: 3
# C:\github\MemeMarketCap\controllers\controller\app.py: 1
# C:\github\MemeMarketCap\controllers\migrations\env.py: 24
Flask == 1.1.1

# C:\github\MemeMarketCap\controllers\controller\extensions.py: 1
Flask_DebugToolbar == 0.10.1

# C:\github\MemeMarketCap\controllers\controller\app.py: 2
Flask_Migrate == 2.5.2

# C:\github\MemeMarketCap\controllers\controller\extensions.py: 2
# C:\github\MemeMarketCap\controllers\controller\generators\sitedata.py: 5
Flask_SQLAlchemy == 2.4.1

# C:\github\MemeMarketCap\controllers\controller\generators\classes\simulator.py: 11
# C:\github\MemeMarketCap\controllers\controller\generators\sitedata.py: 2,3,4
# C:\github\MemeMarketCap\controllers\controller\reddit\classes\redb.py: 2
# C:\github\MemeMarketCap\controllers\controller\reddit\classes\scraper.py: 9
# C:\github\MemeMarketCap\controllers\controller\reddit\functions\database.py: 2
# C:\github\MemeMarketCap\controllers\migrations\env.py: 6,7
# C:\github\MemeMarketCap\controllers\migrations\versions\cfce76273f2a_.py: 9
SQLAlchemy == 1.3.13

# C:\github\MemeMarketCap\controllers\migrations\env.py: 9
# C:\github\MemeMarketCap\controllers\migrations\versions\cfce76273f2a_.py: 8
alembic == 1.4.0

# C:\github\MemeMarketCap\controllers\controller\generators\classes\settings.py: 3
# C:\github\MemeMarketCap\controllers\controller\generators\classes\simulator.py: 6
arrow == 0.15.5

# C:\github\MemeMarketCap\controllers\controller\generators\classes\settings.py: 4
# C:\github\MemeMarketCap\controllers\controller\generators\classes\simulator.py: 7
bcrypt == 3.1.7

# C:\github\MemeMarketCap\controllers\scripts\templates.py: 2
beautifulsoup4 == 4.8.2

# C:\github\MemeMarketCap\controllers\controller\reddit\classes\scraper.py: 8
billiard == 3.6.3.0

# C:\github\MemeMarketCap\controllers\config\flask.py: 3
# C:\github\MemeMarketCap\controllers\controller\app.py: 3
celery == 4.4.0

# C:\github\MemeMarketCap\controllers\controller\generators\classes\simulator.py: 9
# C:\github\MemeMarketCap\controllers\controller\reddit\classes\scraper.py: 10
fabulous == 0.3.0

# C:\github\MemeMarketCap\controllers\controller\reddit\classes\scraper.py: 3
matplotlib == 3.1.3

# C:\github\MemeMarketCap\controllers\controller\generators\classes\simulator.py: 8
# C:\github\MemeMarketCap\controllers\controller\reddit\functions\dataframe.py: 2
numpy == 1.18.1

# C:\github\MemeMarketCap\controllers\controller\reddit\classes\scraper.py: 2
# C:\github\MemeMarketCap\controllers\controller\reddit\functions\dataframe.py: 1
pandas == 0.25.1

# C:\github\MemeMarketCap\controllers\controller\reddit\functions\praw.py: 2
praw == 6.5.1

# C:\github\MemeMarketCap\controllers\config\flask.py: 1
# C:\github\MemeMarketCap\controllers\controller\app.py: 4
# C:\github\MemeMarketCap\controllers\controller\hive\classes\settings.py: 1
# C:\github\MemeMarketCap\controllers\controller\reddit\functions\constants.py: 2
# C:\github\MemeMarketCap\controllers\controller\reddit\functions\praw.py: 4
# C:\github\MemeMarketCap\controllers\controller\sentry.py: 2
python_decouple == 3.3

# C:\github\MemeMarketCap\controllers\controller\reddit\classes\redb.py: 1
redis == 3.4.1

# C:\github\MemeMarketCap\controllers\controller\api\database.py: 1
# C:\github\MemeMarketCap\controllers\controller\reddit\functions\pushshift.py: 1
# C:\github\MemeMarketCap\controllers\scripts\templates.py: 1
requests == 2.22.0

# C:\github\MemeMarketCap\controllers\controller\app.py: 5,6,7,8
# C:\github\MemeMarketCap\controllers\controller\sentry.py: 1
sentry_sdk == 0.14.1

# C:\github\MemeMarketCap\controllers\setup.py: 1
setuptools == 45.1.0

# C:\github\MemeMarketCap\controllers\controller\reddit\classes\scraper.py: 5
# C:\github\MemeMarketCap\controllers\scripts\templates.py: 4
tqdm == 4.42.0

rejson == 0.5.4

# not found by pigar
gunicorn == 20.0.4
psycopg2-binary == 2.8.4
flower == 0.9.3
lxml == 4.5.0
Pillow == 7.1.2

sentry-sdk[flask] == 0.14.1